+++ apis/weatherflow.js
@@ -11,11 +11,12 @@
 const converter = require('../util/converter'),
 	moment = require('moment-timezone'),
 	dgram = require("dgram"),
-	wformula = require('weather-formulas');
+	wformula = require('weather-formulas'),
+	request = require('request');
 
 class TempestAPI
 {
-	constructor (conditionDetail, log, cacheDirectory)
+	constructor (apiKey, locationId, conditionDetail, log, cacheDirectory)
 	{
 		this.attribution = 'Weatherflow Tempest';
 		this.reportCharacteristics = [
@@ -40,6 +41,7 @@ class TempestAPI
 			'LightLevel', // Illuminance
 			'BatteryLevel', // Device Battery level percent
 			'BatteryIsCharging', // Device Battery charging state
+			'StatusFault', // Report if there is a fault
 			
 			// Derived values
 			// @see https://weatherflow.github.io/Tempest/api/derived-metric-formulas.html
@@ -47,9 +49,32 @@ class TempestAPI
 			'TemperatureApparent', // Calculated from Humidity, Temperature & Windspeed
 			'TemperatureWetBulb' // Calculated from Humidity & Temperature
 		];
+		
+		this.forecastCharacteristics = [
+			'ObservationTime', // Forecast update time
+			'Condition', //Conditions
+			'ConditionCategory',
+			'ForecastDay', // day_num, month_num, day_start_local
+			'RainBool', // precip_icon contains === 'rain' || 'sleet' || 'storm'
+			'SnowBool', // precip_icon contains === 'snow'
+			'SunriseTime', //sunrise
+			'SunsetTime', // sunset
+			'TemperatureMax',  // air_temp_high
+			'TemperatureMin', // air_temp_low
+			'RainChance' // precip_probability
+		];
+		this.forecastDays = 10;
+		
+		// Only define the update variable if we have an apiKey and locationId
+		if (apiKey && apiKey.length > 0 && locationId && locationId.length > 0) {
+			this.lastForecastUpdate = -1;
+			}
 	
 		this.conditionDetail = conditionDetail;
 		this.log = log;
+		this.apiKey = apiKey;
+		this.locationId = locationId;
+		
 		this.storage = require('node-persist');
 		// The saved data is only valid for up to 24hrs (TTL)
 		this.storage.initSync({dir:cacheDirectory, forgiveParseErrors: true, ttl: true});
@@ -89,20 +114,27 @@ class TempestAPI
 		this.currentReport.LightningAvgDistance = 0;
 		this.currentReport.LightLevel = 0;
 		this.currentReport.TemperatureWetBulb = 0;
+		this.currentReport.StatusFault = 0;
 
 		// Non-exposed Weather report characteristics
+		// Sky or Tempest station (unlikely to have both)
 		this.currentReport.SkySensorBatteryLevel = 100;
 		this.currentReport.SkySerialNumber = "SK-";
 		this.currentReport.SkyFirmware = "1.0";
+		this.currentReport.SkySensorFailureLog = -1;
+		// Air station
 		this.currentReport.AirSensorBatteryLevel = 100;
 		this.currentReport.AirSerialNumber = "AR-";
 		this.currentReport.AirFirmware = "1.0";
-		this.currentReport.LightLevelSensorFail = 0;
-		this.currentReport.HumiditySensorFail = 0;
-		this.currentReport.TemperatureSensorFail = 0;
+		this.currentReport.AirSensorFailureLog = -1;
+		this.currentReport.SensorString = "Ok";
 
 		// Attempt to restore previous values
 		this.load();
+		
+		// Keep track of previous message so that
+		// we can remove duplicates
+		this.prevMsg = "";
 	
 		// Create UDP listener and start listening
 		this.server = dgram.createSocket({type: 'udp4', reuseAddr: true});
@@ -115,7 +147,12 @@ class TempestAPI
 			try {
 				var message = JSON.parse(msg);
 				this.log.debug(`Server got: ${message.type}`);
-				this.parseMessage(message);
+				if (msg.toString() === this.prevMsg) {
+				    this.log.debug(`Duplicate msg ${msg}`);
+				} else {
+				    this.parseMessage(message);
+				}
+				this.prevMsg = msg.toString();
 				}
 			catch(ex) {
 				this.log(`JSON Parse Exception: ${msg} ${ex}`);
@@ -175,13 +212,40 @@ class TempestAPI
 
 		let weather = {};
 		weather.forecasts = [];
-		let that = this;
-		weather.report = that.currentReport;
-		callback(null, weather);
 		
-		// Save the state after updating plugin state so we don't
-		// delay the update
-		this.save(that.currentReport);
+		// Limit forecast updates to once every hour. Forecast won't change that quickly
+		if ((typeof this.lastForecastUpdate !== 'undefined')  && moment().hour() != this.lastForecastUpdate) {
+			this.lastForecastUpdate = moment().hour();
+			this.getForecastData((error, result) =>
+								 {
+				if (!error) {
+					try {
+						weather.forecasts = this.parseForecasts(result["current_conditions"]["time"], result["forecast"]["daily"], result["timezone"]);
+					} catch (e) {
+						this.log.error("Error parsing weather Forecast");
+						this.log.error(result);
+						this.log.error(e);
+					}
+				} else {
+					this.log.error("Error retrieving weather Forecast");
+					this.log.error(result);
+				}
+				
+				let that = this;
+				weather.report = that.currentReport;
+				callback(null, weather);
+				
+				// Save the state after updating plugin state so we don't
+				// delay the update
+				this.save(that.currentReport);
+			});
+		}
+		else {
+			// Don't update the forecast, just update current conditions
+			let that = this;
+			weather.report = that.currentReport;
+			callback(null, weather);
+		}
 	}
 
 	// Map Tempest precipitation values to Eve Condition Categories
@@ -230,10 +294,11 @@ class TempestAPI
 		return accumulation;
 	}
 
-	// Assume that zero battery level is 2.6v
+	// For AIR/SKY sensor units, assume to be the same as
+	// the Tempest documentation.
 	getBatteryPercent(batteryVoltage)
 	{
-		return (batteryVoltage * 100 - 260);
+		return this.getTempestBatteryPercent(batteryVoltage);
 	}
 	
 	// Tempest battery ranges from 2.355 (low) to 2.8 (full)
@@ -254,23 +319,83 @@ class TempestAPI
 		if (message.type == 'device_status') {
 			that.currentReport.ObservationStation = message.serial_number;
 			that.currentReport.ObservationTime = moment.unix(message.timestamp).format('HH:mm:ss');
-			that.currentReport.TemperatureSensorFail = (message.sensor_status & 0x00000010) ? 1 : 0;
-			that.currentReport.HumiditySensorFail = (message.sensor_status & 0x00000020) ? 1 : 0;
-			that.currentReport.LightLevelSensorFail = (message.sensor_status & 0x00000100) ? 1 : 0;
-			// TODO: Check if a sensor has failed, and log it!
-			this.log.debug("Temperature Sensor Fail: %d, Humidity Sensor Fail: %d, Light Level Sensor Fail: %d", 
-				that.currentReport.TemperatureSensorFail, 
-				that.currentReport.HumiditySensorFail, 
-				that.currentReport.LightLevelSensorFail);
-			var previousLevel = that.currentReport.BatteryLevel;
-			that.currentReport.BatteryIsCharging = false;
-			if (message.serial_number.charAt(1) == 'T') {  // Tempest 
-				that.currentReport.BatteryLevel = this.getTempestBatteryPercent(message.voltage);
+			
+			// Handle sensor failures
+			// Per API v171, only intepret values defined, ignore all others
+			message.sensor_status = message.sensor_status & 0x1FFFF;
+
+			// Any value other than zero for sensor_status means we have a failure
+			that.currentReport.StatusFault = message.sensor_status == 0 ? false : true;
+			if (message.sensor_status == 0) {
+				this.currentReport.SensorString = "Ok";
+				
+				// Reset logging interval for only the unit that is ok
+				// Unit prefixes: AR Air, SK Sky, ST Tempest
+				if (message.serial_number.charAt(1) == 'R') {
+					this.currentReport.AirSensorFailureLog = -1;
+				} else {
+					this.currentReport.SkySensorFailureLog = -1;
+				}
 			} else {
-				that.currentReport.BatteryLevel = this.getBatteryPercent(message.voltage);
-			}
-			if (that.currentReport.BatteryLevel > previousLevel) {
-				that.currentReport.BatteryIsCharging = true;
+				this.currentReport.SensorString = "";
+				if (message.sensor_status & 0x00000001) {
+					this.currentReport.SensorString += "Lightning failed"
+				}
+				if (message.sensor_status & 0x00000002) {
+					if (this.currentReport.SensorString.length > 0) this.currentReport.SensorString += ", ";
+					this.currentReport.SensorString += "Lightning noise"
+				}
+				if (message.sensor_status & 0x00000004) {
+					if (this.currentReport.SensorString.length > 0) this.currentReport.SensorString += ", ";
+					this.currentReport.SensorString += "Lightning disturber"
+				}
+				if (message.sensor_status & 0x00000008) {
+					if (this.currentReport.SensorString.length > 0) this.currentReport.SensorString += ", ";
+					this.currentReport.SensorString += "Pressure failed"
+				}
+				if (message.sensor_status & 0x00000010) {
+					if (this.currentReport.SensorString.length > 0) this.currentReport.SensorString += ", ";
+					this.currentReport.SensorString += "Temperature failed"
+				}
+				if (message.sensor_status & 0x00000020) {
+					if (this.currentReport.SensorString.length > 0) this.currentReport.SensorString += ", ";
+					this.currentReport.SensorString += "Relative Humidity failed"
+				}
+				if (message.sensor_status & 0x00000040) {
+					if (this.currentReport.SensorString.length > 0) this.currentReport.SensorString += ", ";
+					this.currentReport.SensorString += "Wind failed"
+				}
+				if (message.sensor_status & 0x00000080) {
+					if (this.currentReport.SensorString.length > 0) this.currentReport.SensorString += ", ";
+					this.currentReport.SensorString += "Precipitation failed"
+				}
+				if (message.sensor_status & 0x00000100) {
+					if (this.currentReport.SensorString.length > 0) this.currentReport.SensorString += ", ";
+					this.currentReport.SensorString += "Light/UV failed"
+				}
+				if (message.sensor_status & 0x00008000) {
+					if (this.currentReport.SensorString.length > 0) this.currentReport.SensorString += ", ";
+					this.currentReport.SensorString += "Power booster depleted"
+				}
+				if (message.sensor_status & 0x00010000) {
+					if (this.currentReport.SensorString.length > 0) this.currentReport.SensorString += ", ";
+					this.currentReport.SensorString += "Power booster shore power"
+				}
+				this.log.debug("Sensor on unit %s failed error code: %d", message.serial_number, message.sensor_status);
+				
+				// Track error logging per failed device
+				// Unit prefixes: AR Air, SK Sky, ST Tempest
+				if (message.serial_number.charAt(1) == 'R') {
+					if (this.currentReport.AirSensorFailureLog != moment.unix(message.timestamp).hour()) {
+						this.log.error("Sensor on unit %s failed: ", message.serial_number, this.currentReport.SensorString);
+					}
+					this.currentReport.AirSensorFailureLog = moment.unix(message.timestamp).hour();
+				} else {
+					if (this.currentReport.SkySensorFailureLog != moment.unix(message.timestamp).hour()) {
+						this.log.error("Sensor on unit %s failed: ", message.serial_number, this.currentReport.SensorString);
+					}
+					this.currentReport.SkySensorFailureLog = moment.unix(message.timestamp).hour();
+				}
 			}
 		}
 		
@@ -294,31 +419,40 @@ class TempestAPI
 			that.currentReport.AirFirmware = message.firmware_revision;
 			that.currentReport.ObservationTime = moment.unix(message.obs[0][0]).format('HH:mm:ss');
 			that.currentReport.AirPressure = message.obs[0][1];
-			if (that.currentReport.TemperatureSensorFail == 1)
-				that.currentReport.Temperature = 0;
-			else
-				that.currentReport.Temperature = message.obs[0][2];
-			if (that.currentReport.HumiditySensorFail == 1) {
-				that.currentReport.Humidity = 0;
-				that.currentReport.DewPoint = 0;
-			}
-			else {
-				that.currentReport.Humidity = message.obs[0][3];
+			that.currentReport.Temperature = message.obs[0][2];
+			that.currentReport.Humidity = message.obs[0][3];
+			
+			// Only perform new calculations if temperature and humidity values are within a good range
+			// We could get out of range values if the sensors have failed.
+			if (that.currentReport.Humidity > 0 && that.currentReport.Humidity <= 100 &&
+				that.currentReport.Temperature > -100 && that.currentReport.Temperature < 100) {
 				that.currentReport.DewPoint = wformula.kelvinToCelcius(wformula.dewPointMagnusFormula(
 					wformula.celciusToKelvin(that.currentReport.Temperature), 
 					that.currentReport.Humidity));
-			}
-			that.currentReport.TemperatureApparent = wformula.kelvinToCelcius(wformula.australianAapparentTemperature(
+				that.currentReport.TemperatureApparent = wformula.kelvinToCelcius(wformula.australianAapparentTemperature(
 					wformula.celciusToKelvin(that.currentReport.Temperature),
 					that.currentReport.Humidity,
 					that.currentReport.WindSpeed));
+				that.currentReport.TemperatureWetBulb =
+					converter.getWetBulbTemperature(that.currentReport.Temperature, that.currentReport.Humidity);
+			}
 			that.currentReport.TemperatureMin = (that.currentReport.Temperature < that.currentReport.TemperatureMin) ?
 					that.currentReport.Temperature : that.currentReport.TemperatureMin;
-			that.currentReport.TemperatureWetBulb = 
-					converter.getWetBulbTemperature(that.currentReport.Temperature, that.currentReport.Humidity);
 			that.currentReport.LightningStrikes = message.obs[0][4];
 			that.currentReport.LightningAvgDistance = message.obs[0][5];
+			
+			// Report battery status.
+			var previousLevel = that.currentReport.AirSensorBatteryLevel;
 			that.currentReport.AirSensorBatteryLevel = this.getBatteryPercent(message.obs[0][6]);
+			// If the AIR sensor has the lowest battery level, then report it as the Station battery level
+			if (that.currentReport.AirSensorBatteryLevel < that.currentReport.SkySensorBatteryLevel) {
+				that.currentReport.BatteryLevel = that.currentReport.AirSensorBatteryLevel;
+				// It could have a solar panel on it, so check to see if it is going up (charging)
+				that.currentReport.BatteryIsCharging = false;
+				if (that.currentReport.BatteryLevel > previousLevel) {
+					that.currentReport.BatteryIsCharging = true;
+				}
+			}
 		}
 
 		if (message.type == 'obs_sky') {
@@ -326,10 +460,7 @@ class TempestAPI
 			that.currentReport.ObservationStation = that.currentReport.SkySerialNumber;
 			that.currentReport.SkyFirmware = message.firmware_revision;
 			that.currentReport.ObservationTime = moment.unix(message.obs[0][0]).format('HH:mm:ss');
-			if (that.currentReport.LightLevelSensorFail == 1)
-				that.currentReportLightLevel = 0;
-			else
-				that.currentReport.LightLevel = message.obs[0][1];
+			that.currentReport.LightLevel = message.obs[0][1];
 			that.currentReport.UVIndex = message.obs[0][2];
 			that.currentReport.Rain1h = this.getHourlyAccumulatedRain(message.obs[0][0], message.obs[0][3]);
 		
@@ -342,7 +473,19 @@ class TempestAPI
 			that.currentReport.WindSpeedLull = message.obs[0][4];
 			that.currentReport.WindSpeedMax = message.obs[0][6];
 			
+			// Report battery status.
+			var previousLevel = that.currentReport.SkySensorBatteryLevel;
 			that.currentReport.SkySensorBatteryLevel = this.getBatteryPercent(message.obs[0][8]);
+			// If the SKY sensor has the lowest battery level, then report it as the Station battery level
+			if (that.currentReport.SkySensorBatteryLevel < that.currentReport.AirSensorBatteryLevel) {
+				that.currentReport.BatteryLevel = that.currentReport.SkySensorBatteryLevel;
+				// It could have a solar panel on it, so check to see if it is going up (charging)
+				that.currentReport.BatteryIsCharging = false;
+				if (that.currentReport.BatteryLevel > previousLevel) {
+					that.currentReport.BatteryIsCharging = true;
+				}
+			}
+			
 			that.currentReport.SolarRadiation = message.obs[0][10];
 			// Note that Local Day Rain Accumulation (Field 11) is currently always null. Hence we have to approximate it with data available to us.
 			that.currentReport.RainBool = message.obs[0][3] > 0 ? true : false;
@@ -371,30 +514,26 @@ class TempestAPI
             //that.currentReport.WindDirection = converter.getWindDirection(message.obs[0][4]);
             
             that.currentReport.AirPressure = message.obs[0][6];
-	    if (that.currentReport.TemperatureSensorFail == 1)
-	            that.currentReport.Temperature = 0;
-	    else
-	            that.currentReport.Temperature = message.obs[0][7];
-            if (that.currentReport.HumiditySensorFail == 1) {
-                that.currentReport.Humidity = 0;
-                that.currentReport.DewPoint = 0;
-            }
-            else {
-                that.currentReport.Humidity = message.obs[0][8];
+            that.currentReport.Temperature = message.obs[0][7];
+            that.currentReport.Humidity = message.obs[0][8];
+	
+            // Only perform new calculations if temperature and humidity values are within a good range
+            // We could get out of range values if the sensors have failed.
+            if (that.currentReport.Humidity > 0 && that.currentReport.Humidity <= 100 &&
+                    that.currentReport.Temperature > -100 && that.currentReport.Temperature < 100) {
                 that.currentReport.DewPoint = wformula.kelvinToCelcius(wformula.dewPointMagnusFormula(
 					wformula.celciusToKelvin(that.currentReport.Temperature), 
 					that.currentReport.Humidity));
-            }
-	    that.currentReport.TemperatureApparent = wformula.kelvinToCelcius(wformula.australianAapparentTemperature(
+
+                that.currentReport.TemperatureApparent = wformula.kelvinToCelcius(wformula.australianAapparentTemperature(
 					wformula.celciusToKelvin(that.currentReport.Temperature),
 					that.currentReport.Humidity,
 					message.obs[0][2]));
-	    that.currentReport.TemperatureWetBulb = 
+                that.currentReport.TemperatureWetBulb =
 					converter.getWetBulbTemperature(that.currentReport.Temperature, that.currentReport.Humidity);
-	    if (that.currentReport.LightLevelSensorFail == 1)
-                that.currentReport.LightLevel = 0;
-	    else
-                that.currentReport.LightLevel = message.obs[0][9];
+            }
+            
+            that.currentReport.LightLevel = message.obs[0][9];
             that.currentReport.UVIndex = message.obs[0][10];
             that.currentReport.SolarRadiation = message.obs[0][11];
             that.currentReport.Rain1h = this.getHourlyAccumulatedRain(message.obs[0][0], message.obs[0][12]);
@@ -423,14 +562,138 @@ class TempestAPI
             
         }
 	}
-	parseForecasts(forecastObjs, timezone)
+	
+	getForecastConditionCategory(icon, detail = false)
+	{
+		// Convert the icon names to condition category
+		// See https://weatherflow.github.io/Tempest/api/swagger/#!/forecast/getBetterForecast
+		
+		if (icon.includes("thunderstorm") || icon.includes("windy"))
+		{
+			// Severe weather
+			return detail ? 9 : 2;
+		}
+		else if (icon.includes("snow"))
+		{
+			// Snow
+			return detail ? 8 : 3;
+		}
+		else if (icon.includes("sleet"))
+		{
+			// Hail
+			return detail ? 7 : 3;
+		}
+		else if (icon.includes("rain"))
+		{
+			// Rain
+			return detail ? 6 : 2;
+		}
+		else if (icon.includes("fog"))
+		{
+			// Fog
+			return detail ? 4 : 1;
+		}
+		else if (icon.includes("partly-cloudy"))
+		{
+			// Few Clouds
+			return detail ? 1 : 0;
+		}
+		else if (icon.includes("cloudy"))
+		{
+			// Overcast
+			return detail ? 3 : 1;
+		}
+		else if (icon.includes("clear"))
+		{
+			// Clear
+			return 0;
+		}
+		else
+		{
+			this.log.warn("Unknown Tempest Forecast icon " + icon);
+			return 0;
+		}
+	};
+	
+	getForecastData(callback)
 	{
 		/*
-		TODO: Add support for forecasts
 		Weatherflow do provide an API to get forecasts
 		https://weatherflow.github.io/Tempest/api/remote-developer-policy.html
+		https://weatherflow.github.io/Tempest/api/swagger/#!/forecast/getBetterForecast
 		 */
-		return [];
+		this.log.debug("Getting weather forecast for station %s", this.locationId);
+
+		// Response defaults to metric
+		const queryUri = "https://swd.weatherflow.com/swd/rest/better_forecast?station_id=" + this.locationId + "&token=" + this.apiKey;
+		request(encodeURI(queryUri), (requestError, response, body) =>
+		{
+			if (!requestError)
+			{
+				if (response.statusCode == 200) {
+					let parseError;
+					let weather
+					try
+					{
+						weather = JSON.parse(body);
+					} catch (e)
+					{
+						parseError = e;
+					}
+					callback(parseError, weather);
+				} else {
+					callback(true, body);
+				}
+			}
+			else
+			{
+				callback(requestError);
+			}
+		});
+	}
+	
+	
+	parseForecasts(observation_time, values, timezone)
+	{
+		let forecasts = [];
+		 
+		// Check to see if we have 'Todays' forecast as it may be too late
+		// in the day (11pm-midnight) for a daily forecast
+		// If we don't have 'Todays' forecast, then the forecasts will start from tomorrow.
+		let currentDay = moment.unix(observation_time).tz(timezone).date();
+		let forecastOffset = values[0].day_num == currentDay ? 0 : 1;
+		this.log.debug("Weatherflow forecast. Today is:" + currentDay + " First forecast day:" + values[0].day_num);
+		
+		for (let i = 0; i < values.length; i++) {
+			// this.log(values[i]);
+			let forecast = {};
+			forecast.Condition = values[i].conditions;
+			forecast.ConditionCategory = this.getForecastConditionCategory(values[i].icon, this.conditionDetail);
+			forecast.ForecastDay =  moment.unix(values[i].day_start_local).tz(timezone).format('dddd');
+			let detailedCondition = this.getForecastConditionCategory(values[i].icon, true);
+			forecast.RainBool = [5, 6, 9].includes(detailedCondition);
+			forecast.SnowBool = [7, 8].includes(detailedCondition);
+			forecast.SunriseTime = moment.unix(values[i].sunrise).tz(timezone).format('HH:mm:ss');
+			forecast.SunsetTime = moment.unix(values[i].sunset).tz(timezone).format('HH:mm:ss');
+			forecast.TemperatureMax = values[i].air_temp_high;
+			forecast.TemperatureMin = values[i].air_temp_low;
+			forecast.RainChance = values[i].precip_probability;
+			forecast.ObservationTime = moment.unix(observation_time).tz(timezone).format('HH:mm:ss');
+			
+			forecasts[i+forecastOffset] = forecast;
+		}
+		
+		if (forecastOffset == 0) {
+			// If we have the current day forecast save it away
+			this.savedDayForecast = forecasts[0];
+			this.log.debug("Weatherflow storing today's forecast");
+		} else {
+			// If we don't have the current day forecast, present the last saved one
+			this.log.debug("Weatherflow inserting saved today's forecast");
+			forecasts[0] = this.savedDayForecast;
+		}
+		
+		return forecasts;
 	}
 
 }